reorg the export code
